Updated for go 13 and operator sdk 13

+++ internal/pkg/load/validation.go
@@ -14,7 +14,7 @@ import (
 )
 
 // PerformResourceValidation validates templated resources to identify deprecated API versions
-func PerformResourceValidation(resourcesPath string) (*validatemap.ValidateMap, error) {
+func PerformResourceValidation(resourcesPath string,onErrClean bool) (*validatemap.ValidateMap, error) {
 	// collect all templated files in directory
 	// TODO: need to test with nested templates
 	files, err := ioutil.ReadDir(resourcesPath)
@@ -26,40 +26,40 @@ func PerformResourceValidation(resourcesPath string) (*validatemap.ValidateMap, 
 
 	for _, file := range files {
 		rconfig, err := yamlUnmarshalSingleResource(filepath.Join(resourcesPath, file.Name()))
-		if err != nil {
-			reader := bufio.NewReader(os.Stdin)
-			if e := err.Error(); e == "deprecated" {
-				fmt.Printf("Resource: %v has a deprecated API version. Please enter 'Y' to proceed without this resource or 'N' to exit the program: ", reflect.TypeOf(rconfig.resource))
-				c, err := reader.ReadByte()
-				if err != nil {
-					fmt.Println(err)
-					cleanUpAndExit(resourcesPath)
-				}
-				if c == []byte("Y")[0] || c == []byte("y")[0] {
-					addResourceToContinue(&validMap, file.Name())
-				} else {
-					cleanUpAndExit(resourcesPath)
-				}
-			} else if e == "unsupported" {
+			if err != nil {
+				reader := bufio.NewReader(os.Stdin)
+				if e := err.Error(); e == "deprecated" {
+					fmt.Printf("Resource: %v has a deprecated API version. Please enter 'Y' to proceed without this resource or 'N' to exit the program: ", reflect.TypeOf(rconfig.resource))
+					c, err := reader.ReadByte()
+					if err != nil {
+						fmt.Println(err)
+						cleanUpAndExit(resourcesPath,onErrClean)
+					}
+					if c == []byte("Y")[0] || c == []byte("y")[0] {
+						addResourceToContinue(&validMap, file.Name())
+					} else {
+						cleanUpAndExit(resourcesPath,onErrClean)
+					}
+				} else if e == "unsupported" {
 
-				fmt.Printf("Resource: %v is unsupported. Please enter 'Y' to proceed without this resource, or 'N' to exit the program: ", reflect.TypeOf(rconfig.resource))
-				c, err := reader.ReadByte()
-				if err != nil {
-					fmt.Println(err)
-					cleanUpAndExit(resourcesPath)
-				}
-				if c == []byte("Y")[0] || c == []byte("y")[0] {
-					addResourceToContinue(&validMap, file.Name())
-				} else {
-					cleanUpAndExit(resourcesPath)
-				}
+					fmt.Printf("Resource: %v is unsupported. Please enter 'Y' to proceed without this resource, or 'N' to exit the program: ", reflect.TypeOf(rconfig.resource))
+					c, err := reader.ReadByte()
+					if err != nil {
+						fmt.Println(err)
+						cleanUpAndExit(resourcesPath,onErrClean)
+					}
+					if c == []byte("Y")[0] || c == []byte("y")[0] {
+						addResourceToContinue(&validMap, file.Name())
+					} else {
+						cleanUpAndExit(resourcesPath,onErrClean)
+					}
 
-			} else if e == "not yaml" || e == "empty" {
-				continue
-			} else {
-				return nil, fmt.Errorf("uncaught error: %v", err)
+				} else if e == "not yaml" || e == "empty" {
+					continue
+				} else {
+					return nil, fmt.Errorf("uncaught error: %v", err)
+				}
 			}
-		}
 	}
 	return &validMap, nil
 }
@@ -68,17 +68,19 @@ func matchString(text string, contains string) bool {
 	return strings.Contains(text, contains)
 }
 
-func cleanUpAndExit(resourcesPath string) {
+func cleanUpAndExit(resourcesPath string,onErrClean bool) {
 	// log exit begin
-	log.Println("Cleanup Temp Started")
-	// clean up temp folder
-	temp := filepath.Dir(filepath.Dir(resourcesPath))
-	err := os.RemoveAll(temp)
-	if err != nil {
-		log.Printf("Error While Cleaning Up Temp: %v", err)
-	} else {
-		// log cleanup
-		log.Println("Cleanup Temp Completed")
+	if onErrClean{
+		log.Println("Cleanup Temp Started")
+		// clean up temp folder
+		temp := filepath.Dir(filepath.Dir(resourcesPath))
+		err := os.RemoveAll(temp)
+		if err != nil {
+			log.Printf("Error While Cleaning Up Temp: %v", err)
+		} else {
+			// log cleanup
+			log.Println("Cleanup Temp Completed")
+		}
 	}
 	// log exit
 	log.Println("Will Exit")
